Refactors app composables and colors
