+++ app/src/main/java/com/kodeco/android/coordplot/ui/theme/Color.kt
@@ -2,6 +2,7 @@ package com.kodeco.android.coordplot.ui.theme
 
 import androidx.compose.ui.graphics.Color
 
+val Black = Color.Black
 val Purple80 = Color(0xFFD0BCFF)
 val PurpleGrey80 = Color(0xFFCCC2DC)
 val Pink80 = Color(0xFFEFB8C8)
